Fallback to the company payable account if the invoice partner lack it
When generating account moves, we search the term account for the
dedicated invoice line either with:

- An already existing line account
- The partner configured property accounts
- A company dedicated one

The point is that if the partner dont have those accounts correctly
configured, then no account is used which lead to the issue:

`account_move_line_check_accountable_required_fields`

which relate to the SQL constraint :

"check_accountable_required_fields",
"CHECK(COALESCE(display_type IN ('line_section', 'line_note'), 'f')
	OR account_id IS NOT NULL)",
"Missing required account on accountable invoice line."

Here this is the down payment line who cause this as it have False as
display_type and NULL as account_id.

To solves this, we will fallback to the next way to search the
account_id and raise a dedicated issue if no account is found. The issue
will help the user to figure out how to solve the case.