Fix an issue generating invoices on  multi-currencies setups
When looking for the right account when generating account moves, we
have to take care to use the currency corresponding account.

In the current state, the code will fail later when checking constraint
cause some line will use one currency and the balande line another.

This add the account move currency_id to the account domain to make sure
we pickup the corresponding one.